// packages/shared/src/emotion/emotion.schema.d.ts
export declare const CoreMoodSchema: z.ZodObject<{
    valence: z.ZodNumber;
    arousal: z.ZodNumber;
    stability: z.ZodNumber;
    energy: z.ZodNumber;
    circumplexAngle: z.ZodNumber;
    circumplexRadius: z.ZodNumber;
    momentumValence: z.ZodNumber;
    momentumArousal: z.ZodNumber;
}, "strip", z.ZodTypeAny, {
    valence: number;
    arousal: number;
    stability: number;
    energy: number;
    circumplexAngle: number;
    circumplexRadius: number;
    momentumValence: number;
    momentumArousal: number;
}, {
    valence: number;
    arousal: number;
    stability: number;
    energy: number;
    circumplexAngle: number;
    circumplexRadius: number;
    momentumValence: number;
    momentumArousal: number;
}>;
export declare const SocialEmotionsSchema: z.ZodObject<{
    jealousy: z.ZodNumber;
    envy: z.ZodNumber;
    humiliation: z.ZodNumber;
    pride: z.ZodNumber;
    shame: z.ZodNumber;
    affection: z.ZodNumber;
    resentment: z.ZodNumber;
    suspicion: z.ZodNumber;
    admiration: z.ZodNumber;
    contempt: z.ZodNumber;
    neediness: z.ZodNumber;
    socialAnxiety: z.ZodNumber;
    fearOfExclusion: z.ZodNumber;
    desireForStatus: z.ZodNumber;
    desireForIntimacy: z.ZodNumber;
}, "strip", z.ZodTypeAny, {
    jealousy: number;
    envy: number;
    humiliation: number;
    pride: number;
    shame: number;
    affection: number;
    resentment: number;
    suspicion: number;
    admiration: number;
    contempt: number;
    neediness: number;
    socialAnxiety: number;
    fearOfExclusion: number;
    desireForStatus: number;
    desireForIntimacy: number;
}, {
    jealousy: number;
    envy: number;
    humiliation: number;
    pride: number;
    shame: number;
    affection: number;
    resentment: number;
    suspicion: number;
    admiration: number;
    contempt: number;
    neediness: number;
    socialAnxiety: number;
    fearOfExclusion: number;
    desireForStatus: number;
    desireForIntimacy: number;
}>;